+++ ts/features/pn/aar/components/SendQRScanRedirectComponent.tsx
