+++ ts/features/pn/aar/components/__mocks__/SendQRScanRedirectComponent.tsx
