+++ ts/features/pn/aar/components/__tests__/SendQrScanRedirectComponent.test.tsx
